requirements.txt
#Django – основной фреймворк.
#psycopg2-binary – драйвер для PostgreSQL. Если вы используете только SQLite по умолчанию, это можно не устанавливать. Но пусть остается — может пригодиться, если в будущем захотите перейти на PostgreSQL.
#Pillow – для работы с изображениями (аватары пользователей, вложения к заметкам).
#python-dotenv – для загрузки переменных окружения из .env.
#djanpython manage.py migratego-crispy-forms и django-crispy-bootstrap5 – для удобного и красивого рендеринга форм с использованием Bootstrap 5.
#django-taggit – для добавления функционала тегов к заметкам и задачам.
Django==4.2
psycopg2-binary==2.9.6
Pillow==9.5.0
python-dotenv==1.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-taggit==4.0.0

